requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
jinja2==3.1.2
pillow==10.1.0
numpy==1.26.2
librosa==0.10.1
trimesh==4.0.5
nltk==3.8.1
pandas==2.1.3
soundfile==0.12.1
scipy==1.11.3 